feat(system): fix(prax): reach 100% seedgo — inline monitor_info.py into monitor.py, eliminate CLI separation violation (DPLAN-0141)
