+++ src/aipass/api/apps/modules/bridge.py
@@ -59,3 +59,11 @@ def list_contracts() -> list[str]:
 def clear() -> None:
     """Clear all registrations. Intended for test teardown only."""
     _registry.clear()
+
+
+def handle_command(command: str, args: list) -> bool:
+    """Bridge is a utility module — no drone commands. Always returns False."""
+    if not args:
+        print_introspection()
+        return False
+    return False
