+++ src/aipass/api/tests/test_integrations.py
@@ -3,22 +3,24 @@
 # Description: Tests for bridge, registry, and integrations handlers
 # Version: 1.0.0
 # Created: 2026-04-15
-# Modified: 2026-04-15
+# Modified: 2026-04-22
 # =============================================
 """
 Tests for DPLAN-0133 Phase 2: bridge + registry + handlers.
 
 Groups:
-  TestBridge                — contract registration, resolve, list, clear
-  TestRegistry              — auto-discovery walk, empty dir, missing driver, broken import
-  TestFetchContracts        — fetch_contracts() happy path and empty
-  TestCallContract          — call_contract() happy path, unregistered, args forwarding, exception
+  TestBridge            — contract registration, resolve, list, clear
+  TestRegistry          — auto-discovery walk, empty dir, missing driver, broken import
+  TestListHandler       — get_contracts() happy path and empty
+  TestCallHandler       — invoke() happy path, args forwarding, exception handling
 """
 
 import pytest
 
 from aipass.api.apps.modules import bridge, registry
-from aipass.api.apps.modules.integrations_manager import fetch_contracts, call_contract
+from aipass.api.apps.modules.bridge import list_contracts, resolve
+from aipass.api.apps.handlers.integrations.list import get_contracts
+from aipass.api.apps.handlers.integrations.call import invoke
 
 
 # ---------------------------------------------------------------------------
@@ -27,13 +29,13 @@ from aipass.api.apps.modules.integrations_manager import fetch_contracts, call_c
 
 
 @pytest.fixture(autouse=True)
-def clean_bridge():
-    """Reset bridge state before and after each test."""
+def clean_bridge(monkeypatch):
+    """Reset bridge and registry state before and after each test."""
     bridge.clear()
-    registry.reset()
+    monkeypatch.setattr(registry, "_loaded", False)
     yield
     bridge.clear()
-    registry.reset()
+    monkeypatch.setattr(registry, "_loaded", False)
 
 
 # ---------------------------------------------------------------------------
@@ -90,7 +92,7 @@ class TestBridge:
 
 class TestRegistry:
     def test_load_drivers_empty_dir(self, tmp_path):
-        """Empty integrations dir → 0 drivers, no error."""
+        """Empty integrations dir returns 0 drivers, no error."""
         count = registry.load_drivers(integrations_dir=tmp_path)
         assert count == 0
 
@@ -136,7 +138,7 @@ class TestRegistry:
         assert bridge.resolve("goodcontract") is not None
 
     def test_load_drivers_nonexistent_dir(self, tmp_path):
-        """Non-existent integrations dir → 0, no error."""
+        """Non-existent integrations dir returns 0, no error."""
         count = registry.load_drivers(integrations_dir=tmp_path / "nope")
         assert count == 0
 
@@ -150,50 +152,48 @@ class TestRegistry:
 
 
 # ---------------------------------------------------------------------------
-# TestFetchContracts
+# TestListHandler
 # ---------------------------------------------------------------------------
 
 
-class TestFetchContracts:
+class TestListHandler:
     def test_empty_returns_success(self):
-        """fetch_contracts() returns success with empty list when bridge is clear."""
-        result = fetch_contracts()
+        """get_contracts() returns success with empty list when bridge is clear."""
+        result = get_contracts(list_contracts())
         assert result["success"] is True
         assert result["contracts"] == []
         assert result["count"] == 0
 
     def test_returns_registered_contracts(self):
-        """fetch_contracts() returns sorted contracts from bridge."""
+        """get_contracts() returns sorted contracts from bridge."""
         bridge.register("beta", lambda: None)
         bridge.register("alpha", lambda: None)
-        result = fetch_contracts()
+        result = get_contracts(list_contracts())
         assert result["success"] is True
         assert result["contracts"] == ["alpha", "beta"]
         assert result["count"] == 2
 
 
 # ---------------------------------------------------------------------------
-# TestCallContract
+# TestCallHandler
 # ---------------------------------------------------------------------------
 
 
-class TestCallContract:
-    def test_call_registered_contract(self):
-        """call_contract() resolves and invokes registered driver."""
+class TestCallHandler:
+    def test_invoke_registered_contract(self):
+        """invoke() calls registered driver and returns success."""
         bridge.register("ping", lambda *a: "pong")
-        result = call_contract("ping", [])
+        driver_fn = resolve("ping")
+        result = invoke(driver_fn, "ping", [])
         assert result["success"] is True
         assert result["result"] == "pong"
 
-    def test_call_unregistered_returns_failure(self):
-        """call_contract() returns failure for unregistered contract."""
-        result = call_contract("nope", [])
-        assert result["success"] is False
-        assert result["error"] is not None
-        assert "nope" in result["error"]
+    def test_unregistered_contract_resolves_none(self):
+        """resolve() returns None for unregistered contract."""
+        assert resolve("nope") is None
 
-    def test_call_passes_args_to_driver(self):
-        """call_contract() forwards args to the driver function."""
+    def test_invoke_passes_args_to_driver(self):
+        """invoke() forwards args to the driver function."""
         received: list = []
 
         def capturing_driver(*args):
@@ -202,19 +202,21 @@ class TestCallContract:
             return "done"
 
         bridge.register("cap", capturing_driver)
-        result = call_contract("cap", ["foo", "bar"])
+        driver_fn = resolve("cap")
+        result = invoke(driver_fn, "cap", ["foo", "bar"])
         assert result["success"] is True
         assert received == ["foo", "bar"]
 
-    def test_call_driver_exception_returns_failure(self):
-        """call_contract() returns failure dict when driver raises."""
+    def test_invoke_driver_exception_returns_failure(self):
+        """invoke() returns failure dict when driver raises."""
 
         def broken_driver(*args):
             """Always raises to simulate a broken driver."""
             raise RuntimeError("boom")
 
         bridge.register("broken", broken_driver)
-        result = call_contract("broken", [])
+        driver_fn = resolve("broken")
+        result = invoke(driver_fn, "broken", [])
         assert result["success"] is False
         assert result["error"] is not None
         assert "boom" in result["error"]
